Make locator_class completely dynamic (for cell/row)
