+++ lib/watir-webdriver.rb
@@ -14,8 +14,8 @@ require 'watir-webdriver/locators/element_locator'
 require 'watir-webdriver/locators/button_locator'
 require 'watir-webdriver/locators/text_area_locator'
 require 'watir-webdriver/locators/text_field_locator'
-require 'watir-webdriver/locators/child_row_locator'
-require 'watir-webdriver/locators/child_cell_locator'
+require 'watir-webdriver/locators/row_locator'
+require 'watir-webdriver/locators/cell_locator'
 require 'watir-webdriver/browser'
 require 'watir-webdriver/screenshot'
 require 'watir-webdriver/after_hooks'
@@ -93,6 +93,7 @@ require 'watir-webdriver/elements/svg_elements'
 
 require 'watir-webdriver/elements/area'
 require 'watir-webdriver/elements/button'
+require 'watir-webdriver/elements/cell'
 require 'watir-webdriver/elements/checkbox'
 require 'watir-webdriver/elements/dlist'
 require 'watir-webdriver/elements/file_field'
@@ -105,6 +106,7 @@ require 'watir-webdriver/elements/input'
 require 'watir-webdriver/elements/link'
 require 'watir-webdriver/elements/option'
 require 'watir-webdriver/elements/radio'
+require 'watir-webdriver/elements/row'
 require 'watir-webdriver/elements/select'
 require 'watir-webdriver/elements/table'
 require 'watir-webdriver/elements/table_cell'
