+++ lib/watir-webdriver/locators/cell_locator.rb
@@ -1,0 +1,7 @@
+module Watir
+  class CellLocator < ElementLocator
+  end
+end
+
+require 'watir-webdriver/locators/finders/cell_finder'
+require 'watir-webdriver/locators/selector_builders/cell_selector_builder'
