--- lib/watir-webdriver/locators/child_cell_locator.rb
@@ -1,7 +1,0 @@
-module Watir
-  class ChildCellLocator < ElementLocator
-  end
-end
-
-require 'watir-webdriver/locators/finders/child_cell_finder'
-require 'watir-webdriver/locators/selector_builders/child_cell_selector_builder'
