--- lib/watir-webdriver/locators/child_row_locator.rb
@@ -1,7 +1,0 @@
-module Watir
-  class ChildRowLocator < ElementLocator
-  end
-end
-
-require 'watir-webdriver/locators/finders/child_row_finder'
-require 'watir-webdriver/locators/selector_builders/child_row_selector_builder'
