+++ lib/watir-webdriver/locators/row_locator.rb
@@ -1,0 +1,7 @@
+module Watir
+  class RowLocator < ElementLocator
+  end
+end
+
+require 'watir-webdriver/locators/finders/row_finder'
+require 'watir-webdriver/locators/selector_builders/row_selector_builder'
