+++ lib/watir-webdriver/locators/selector_builders/row_selector_builder.rb
@@ -1,10 +1,15 @@
 module Watir
-  class ChildCellLocator
+  class RowLocator
     class SelectorBuilder < ElementLocator::SelectorBuilder
       def build_wd_selector(selectors)
         return if selectors.values.any? { |e| e.kind_of? Regexp }
+        selectors.delete(:tag_name) || raise("internal error: no tag_name?!")
 
-        expressions = %w[./th ./td]
+        expressions = %w[./tr]
+        unless %w[tbody tfoot thead].include?(@wd.tag_name.downcase)
+          expressions += %w[./tbody/tr ./thead/tr ./tfoot/tr]
+        end
+
         attr_expr = attribute_expression(selectors)
 
         unless attr_expr.empty?
