+++ lib/watir-webdriver/locators/text_field_locator.rb
@@ -1,6 +1,5 @@
 module Watir
   class TextFieldLocator < ElementLocator
-    NON_TEXT_TYPES = %w[file radio checkbox submit reset image button hidden datetime date month week time datetime-local range color]
   end # TextFieldLocator
 end # Watir
